[sourcegen] Make sourcegen a CLI
